Now runs and displays player entity.
Fixed various issues preventing game from running.
Created player object and drawing it on the screen.

+++ ProjectPinetree/Source/sprite.cpp
@@ -41,8 +41,8 @@ Sprite::Initialise(Texture& texture)
 void 
 Sprite::Process(float deltaTime)
 {
-	m_x += m_dx * deltaTime;
-	m_y += m_dy * deltaTime;
+	m_x += (int)((float)m_dx * deltaTime);
+	m_y += (int)((float)m_dy * deltaTime);
 }
 
 void
